[fs] Move generic filesystem code to its own module
